Make page size a server config option via env var
The number of records to ask Flickr for in each request should be a tunable
param.
We want a balance of perf and responsiveness.
Too many results in each request may result in Tableau becoming unresponsive.
Too few increases to the total time to get all the results.
Start with 100 as the default for now, tunable via an environment variable:
FLICKR_WDC_PAGESIZE.

Resolves #15.
The max Flickr supports is 500.