Check cookie as part of determining when to auth
We expect the oauth_token query params only when being redirected after oauth
sign-in.
Turns out that Tableau saves the complete URL and uses that when it's time to
refresh the extract.
To work around this, use the cookie that the proxy sets just before oauth
sign-in as part of the condition to use when deciding if this is a redirect.
Maybe it could be the only condidtion?