Create generic /forward method on server
Instead of creating a special server method for each type of Flickr API call,
create a generic /forward that sets common params and then signs the request and
forwards it along to Flickr.
We lose a little customized server-side logging with this and the ability to
configure the page size and a debugging page limit on the server this way.
That's probably OK.

Closes #18.